Added dt to MDPInfo
- now MDPInfo has a field dt representing the control timestep
- all environment set correctly or using a default value dt
- removed _dt variables in the mushroom environments
- code and comments refactoring
- fixed docstring in viewer

+++ mushroom_rl/utils/viewer.py
@@ -78,8 +78,8 @@ class Viewer:
         Constructor.
 
         Args:
-            env_width (int): The x dimension limit of the desired environment;
-            env_height (int): The y dimension limit of the desired environment;
+            env_width (float): The x dimension limit of the desired environment;
+            env_height (float): The y dimension limit of the desired environment;
             width (int, 500): width of the environment window;
             height (int, 500): height of the environment window;
             background (tuple, (0, 0, 0)): background color of the screen.
